done some work

+++ internal/api/handlers/getbalance.go
@@ -8,8 +8,8 @@ import (
 )
 
 type GetBalanceResponseData struct {
-	Current   float64 `json:"current"`
-	Withdrawn float64 `json:"withdrawn"`
+	Current   float32 `json:"current"`
+	Withdrawn float32 `json:"withdrawn"`
 }
 
 func (h Handlers) HandleGetBalance(w http.ResponseWriter, r *http.Request) {
